Fix CICE restart fail - freeing iodesc
Closing the File handle after we free the associated io
descriptors. Without the change when freeing io descriptors PIO
tries to perform a syncfile on the File handle (that is no longer
valid because it was closed) resulting in an abort (PIO sync fails
since it cannot find the file handle).