Add smoothing flag to cube_to_target build
Add a preprocessor flag (SMOOTH_TERR) to cube_to_target to enable
building with terrain smoothing enabled. This is kind of an inelegant
way of doing (better would be to write up an interface to pass command
line arguments to cube_to_target), but is slightly better than requiring
editing the source code to change the smoothing flags. To compile with
smoothing turned on then, just export FFLAGS="-DSMOOTH_TERR", and then
execute make.