Expose namelist option cubed_sphere_map
Expose namelist option cubed_sphere_map for SE dycore. This specifies the type
of mapping from the reference element space to the sphere, and is required for
the new TempestRemap mappings and inputdata based on those maps due to needing
to use a different projection than was default in V1.