Make get_cloud_optics routines consistent
Make interfaces to get_cloud_optics routines consistent. This just involved
adding the `ncol` argument to those that did not already have it to specify how
many columns to work on (in the event that `ncol < pcols`). The alternative is
to get this from something like `ncol = size(cld_liq_path)`, but the advantage
to specifying the dimension size is that we avoid needing to create array
temporaries in the event we want to work over `ncol < pcols` (because we do not
need to resize the inputs in this case).