Simplify underlying optics calling structure
Remove redundant subroutines and simplify the low-level optics routines to
improve transparency of the code and open up the possibility of exposing more
parallelism in the future. In the process, enforce more consistent variable
naming for routines in cloud_rad_props (i.e., keep the `lambda` and `mu` names,
as opposed to renaming these to `slope_param` and `shape_param` in the
user-facing interface so that the code is internally consistent with the look-up
tables that are read in, and with the variable naming in the literature).