Move extra level handling out of optics routines
Move the handling of the extra level above model top (for radiative heating) out
of the individual optics routines and instead handle with a new subroutine. This
simplifies each routine, and also makes it a little more clear what each routine
is responsible for and what the code is doing.