Make build-namelist write to stderr upon fail
Previously, when a namelist variable was encountered for which a value
failed to be set, the build-namelist script would die but print the
error message using the `print` statement rather than with `die`. The
result was that the error message would be written to stdout, rather
than stderr, but output from build-namelist written to stdout is not
written anywhere when run within CIME. So making sure these get written
to stderr makes sure we get a useful error message.